fix test failure due to differing u64 alignment on different targets

+++ tests/ui/layout/enum.rs
@@ -19,6 +19,6 @@ enum UninhabitedVariantSpace { //~ERROR: size: Size(16 bytes)
 
 #[rustc_layout(abi)]
 enum ScalarPairDifferingSign { //~ERROR: abi: ScalarPair
-    A(u64),
-    B(i64),
+    A(u8),
+    B(i8),
 }
